use zig as build system

+++ code/client/cl_cgame.c
@@ -414,7 +414,7 @@ The cgame module is making a system call
 */
 #define	VMA(x) VM_ArgPtr(args[x])
 #define	VMF(x)	((float *)args)[x]
-int CL_CgameSystemCalls( int *args ) {
+int CL_CgameSystemCalls( intptr_t *args ) {
 	switch( args[0] ) {
 	case CG_PRINT:
 		Com_Printf( "%s", VMA(1) );
