+++ code/server/sv_game.c
@@ -314,7 +314,8 @@ The module is making a system call
 
 #define	VMF(x)	((float *)args)[x]
 
-int SV_GameSystemCalls( int *args ) {
+int SV_GameSystemCalls( intptr_t *args ) {
+
 	switch( args[0] ) {
 	case G_PRINT:
 		Com_Printf( "%s", VMA(1) );
